+++ lib/mix/lib/mix/tasks/deps.clean.ex
@@ -1,23 +1,31 @@
 defmodule Mix.Tasks.Deps.Clean do
   use Mix.Task
 
-  @shortdoc "Remove dependencies' files"
+  @shortdoc "Remove the given dependencies' files"
   @recursive :both
 
   @moduledoc """
-  Clean dependencies.
+  Remove the given dependencies' files.
 
-  By default, cleans all dependencies. A list of dependencies can
-  be given to clean specific ones. Clean does not unlock
-  the repositories, unless `--unlock` is given.
+  Since this is a destructive action, cleaning of all dependencies
+  can only happen by passing the `--all` command line option.
+
+  Clean does not unlock the repositories, unless `--unlock` is given.
   """
 
   import Mix.Deps, only: [all: 0, by_name: 1, format_dep: 1]
 
   def run(args) do
-    case OptionParser.parse(args, switches: [unlock: :boolean]) do
-      { opts, [] }   -> do_clean all, opts
-      { opts, args } -> do_clean by_name(args), opts
+    { opts, args } = OptionParser.parse(args, switches: [unlock: :boolean, all: :boolean])
+
+    cond do
+      opts[:all] ->
+        do_clean all, opts
+      args != [] ->
+        do_clean by_name(args), opts
+      true ->
+        raise Mix.Error, message: "mix deps.clean expects dependencies as arguments or " <>
+                                  "the --all option to clean all dependencies"
     end
   end